requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
scikit-learn
python-multipart
